major frontend code reorgs, expect broken things
