+++ includes/frontend/frontend.hpp
@@ -7,7 +7,7 @@
 
 class Frontend {
 public:
-  explicit Frontend() : gbc_(std::make_unique<GameBoyColor>(*this)) {}
+  explicit Frontend() : gbc(std::make_unique<GameBoyColor>(*this)) {}
   virtual std::array<std::uint32_t, 160 * 144> get_frame() = 0;
   virtual void put_pixel(int x, int y, std::uint32_t c) = 0;
   virtual void clear(std::uint32_t c = 0x00FFFFFF) = 0;
@@ -15,10 +15,10 @@ public:
                                  std::size_t sample_count) = 0;
   virtual void start() = 0;
 
-  std::unique_ptr<GameBoyColor> &get() { return gbc_; }
+  std::unique_ptr<GameBoyColor> &get() { return gbc; }
 
 protected:
-  std::unique_ptr<GameBoyColor> gbc_;
+  std::unique_ptr<GameBoyColor> gbc;
 };
 
 #endif // __FRONTEND_H
