+++ main.cpp
@@ -1,4 +1,4 @@
-#include "frontend/sdl3_frontend.hpp"
+#include "frontend/sdl3/frontend.hpp"
 
 int main(const int argc, const char **argv) {
   auto emulator = SDL3Frontend();
